* Create a `TranslationViewController` to use as a sub view controller in `SearchViewController` * `TranslationViewController` will be easy to reuse to display favorite detail
