+++ weTranslate/SearchViewController.swift
@@ -15,43 +15,11 @@ protocol SearchViewControllerDelegate: class {
 
 final class SearchViewController: UIViewController {
 
-    // MARK: - Type
-
-    enum State {
-        case Default
-        case Loading
-        case Result(TranslationViewModel)
-        case NoResult(NoResultViewModel)
-        case Error
-    }
-
-
     // MARK: - Properties
 
-    var state: State = .Default {
-        didSet {
-            switch state {
-            case .Default:
-                noResultView.hidden = true
-                tableView.hidden = false
-                viewModel = nil
-            case .Loading:
-                noResultView.hidden = true
-                tableView.hidden = false
-            case .Result(let translationViewModel):
-                noResultView.hidden = true
-                tableView.hidden = false
-                viewModel = translationViewModel
-            case .NoResult(let noResultViewModel):
-                noResultView.hidden = false
-                tableView.hidden = true
-                noResultView.viewModel = noResultViewModel
-            case .Error:
-                noResultView.hidden = false
-                tableView.hidden = true
-            }
-            tableView.reloadData()
-        }
+    var state: TranslationViewController.State {
+        get { return translationViewController.state }
+        set { translationViewController.state = newValue }
     }
 
     private var viewModel: TranslationViewModel?
@@ -59,37 +27,19 @@ final class SearchViewController: UIViewController {
     private weak var delegate: SearchViewControllerDelegate?
 
     private let searchHeaderViewController: SearchHeaderViewController = {
-        let searchHeaderViewController = SearchHeaderViewController()
-        return searchHeaderViewController
+        let viewController = SearchHeaderViewController()
+        return viewController
     }()
 
-    private let tableView: UITableView = {
-        let tableView = UITableView()
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        tableView.estimatedRowHeight = 70
-        tableView.rowHeight = UITableViewAutomaticDimension
-        tableView.separatorStyle = .None
-        return tableView
-    }()
-
-    private let translationView: TranslationView = {
-        let translationView = TranslationView(frame: .zero)
-        translationView.translatesAutoresizingMaskIntoConstraints = false
-        return translationView
-    }()
-
-    private let noResultView: NoResultView = {
-        let view = NoResultView(frame: .zero)
-        view.translatesAutoresizingMaskIntoConstraints = false
-        view.hidden = true
-        return view
+    private let translationViewController: TranslationViewController = {
+        let viewController = TranslationViewController()
+        return viewController
     }()
 
     private let bodyView: UIStackView = {
         let stackView = UIStackView()
         stackView.translatesAutoresizingMaskIntoConstraints = false
         stackView.axis = .Vertical
-        stackView.spacing = 5
         stackView.opaque = true
         stackView.layoutMargins = UIEdgeInsets(top: 20, left: 0, bottom: 0, right: 0)
         stackView.layoutMarginsRelativeArrangement = true
@@ -121,15 +71,12 @@ final class SearchViewController: UIViewController {
         searchHeaderViewController.didMoveToParentViewController(self)
         searchHeaderViewController.delegate = self
 
-        tableView.registerClass(WordTableViewCell.self, forCellReuseIdentifier: WordTableViewCell.cellIdentifier)
-        tableView.registerClass(TableViewCell.self, forCellReuseIdentifier: TableViewCell.cellIdentifier)
-        tableView.registerClass(LoadingTableViewCell.self, forCellReuseIdentifier: LoadingTableViewCell.cellIdentifier)
-        tableView.delegate = self
-        tableView.dataSource = self
+        addChildViewController(translationViewController)
+        translationViewController.didMoveToParentViewController(self)
+        translationViewController.delegate = self
 
         bodyView.addArrangedSubview(searchHeaderViewController.view)
-        bodyView.addArrangedSubview(noResultView)
-        bodyView.addArrangedSubview(tableView)
+        bodyView.addArrangedSubview(translationViewController.view)
         view.addSubview(bodyView)
 
         bodyView.leadingAnchor.constraintEqualToAnchor(view.leadingAnchor).active = true
@@ -149,72 +96,6 @@ final class SearchViewController: UIViewController {
     }
 }
 
-extension SearchViewController: UITableViewDataSource {
-
-    func numberOfSectionsInTableView(tableView: UITableView) -> Int {
-        guard case .Result(let translationViewModel) = state else { return 1 }
-
-        return translationViewModel.translation.meanings.count
-    }
-
-    func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        guard case .Result(let translationViewModel) = state else { return 1 }
-
-        return translationViewModel.translation.meanings[section].translatedWords.count
-    }
-
-    func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
-
-        switch state {
-        case .Default:
-            break
-        case .Loading:
-            if let cell = tableView.dequeueReusableCellWithIdentifier(LoadingTableViewCell.cellIdentifier, forIndexPath: indexPath) as? LoadingTableViewCell {
-                cell.activityIndicator.startAnimating()
-                return cell
-            }
-        case .Result(_):
-            break
-        case .NoResult:
-            break
-        case .Error:
-            break
-        }
-
-        guard case .Result(let translationViewModel) = state else { return UITableViewCell() }
-
-
-        let wordViewModel = WordViewModel(word: translationViewModel.translation.meanings[indexPath.section].translatedWords[indexPath.row])
-
-        if indexPath.row == 0  && indexPath.section == 0 {
-            translationView.viewModel = wordViewModel
-            return TableViewCell(reuseIdentifier: "TranslationTableViewCell", view: translationView)
-
-        } else if let cell = tableView.dequeueReusableCellWithIdentifier(WordTableViewCell.cellIdentifier, forIndexPath: indexPath) as? WordTableViewCell {
-            cell.viewModel = wordViewModel
-            return cell
-        }
-
-        return UITableViewCell()
-    }
-}
-
-
-extension SearchViewController: UITableViewDelegate {
-
-    func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
-        // FIXME: Push to detail
-    }
-}
-
-
-extension SearchViewController: UIScrollViewDelegate {
-
-    func scrollViewWillBeginDragging(scrollView: UIScrollView) {
-        searchHeaderViewController.resignFirstResponder()
-    }
-}
-
 
 extension SearchViewController: SearchHeaderViewControllerDelegate {
     func searchHeaderViewController(searchHeaderViewController: SearchHeaderViewController, didSearchWord word: String, fromLanguage: Language, toLanguage: Language) {
@@ -225,5 +106,11 @@ extension SearchViewController: SearchHeaderViewControllerDelegate {
     func searchHeaderViewControllerDidResetSearch(searchHeaderViewController: SearchHeaderViewController) {
         state = .Default
     }
+}
 
+
+extension SearchViewController: TranslationViewControllerDelegate {
+    func translationViewControllerBecomeFirstResponder(translationViewController: TranslationViewController) {
+        searchHeaderViewController.resignFirstResponder()
+    }
 }
